FindProcWithLib now returns the list of softerrors
There was another approach to solving this issue:
Returning only one error which the caller should check if it is a
harderror or a softerror (something similar to PathError in OS Package),
but the problem with that is that softerrors are not errors per se: The
function may have succeded for a lot of processes and failed only for a
few ones, thus, the returning value is still useful. The logic for
handling that case would be a bit complicated (instead of: If it is not a
hard error, then I have useful results).

This resolves #8.