Fix race condition on TestSearchInOtherProcess
There was a race condition between when we spawn the process and when the process started running: if we searched the memory before the child process initialized its own memory, we may fail our test.

The quick and dirty solution was to synchronize the processes via stdout: From the go test we wait until the process writes something to its StdoutPipe, and from the C process we call printf and fflush after initializing all the memory.