[memacess/linux] Read the entire process memory
Until we have a good explanation and a different API giving the user the
ability to choose what to read, we read the entire memory of a process.

Also, the previous implementation was hiding several errors in the code, so
it's better if we cover all the cases now, and in the future give add
more granularity about what to read.